DB schema simplification

+++ index.js
@@ -41,7 +41,6 @@ const emptyDir = require('./modules/emptyDir');
 const notify = require('./modules/notify');
 const existPath = require('./modules/existPath');
 
-const createAssignmentsTable = require('./modules/createAssignmentsTable');
 const createLocationsTable = require('./modules/createLocationsTable');
 const createTasksTable = require('./modules/createTasksTable');
 const createUsersTable = require('./modules/createUsersTable');
@@ -58,7 +57,6 @@ const createDB = (config, DBname = process.env.DB) => {
         // Tables have to be created in this exact order to avoid errors when assigning foreign key constraints
         createLocationsTable(client, process.env.LOCATIONS.split(','));
         createUsersTable(client);
-        createAssignmentsTable(client);
         createTasksTable(client);
         // .then(() => {
         //   client.query({
@@ -198,169 +196,169 @@ initClient.connect()
 app.use("/src", express.static(__dirname + "/src"));
 
 app.get('/', (req, res) => {
-    res.render('index.ejs', {
-      currentVersion: tag,
-      isSearchPage: false,
-      locations: process.env.LOCATIONS
-    });
+  res.render('index.ejs', {
+    currentVersion: tag,
+    isSearchPage: false,
+    locations: process.env.LOCATIONS
+  });
 
-    io.once('connection', io => {
-      // const updateBarcode = () => {
-      //   DBquery(io, 'SELECT', 'barcodes', {
-      //       text: `SELECT barcode, available FROM barcodes WHERE available = true LIMIT 1`
-      //     })
-      //     .then(res => {
-      //       let code = res.rows[0].barcode;
-      //
-      //       io.emit('barcode', code);
-      //     })
-      //     .catch(err => {
-      //       notify(io, 'barcode');
-      //       console.error(`Une erreur est survenue lors de l'attribution d'un numéro d'exemplaire : ${err}`);
-      //     });
-      // }
-      // updateBarcode();
+  io.once('connection', io => {
+    // const updateBarcode = () => {
+    //   DBquery(io, 'SELECT', 'barcodes', {
+    //       text: `SELECT barcode, available FROM barcodes WHERE available = true LIMIT 1`
+    //     })
+    //     .then(res => {
+    //       let code = res.rows[0].barcode;
+    //
+    //       io.emit('barcode', code);
+    //     })
+    //     .catch(err => {
+    //       notify(io, 'barcode');
+    //       console.error(`Une erreur est survenue lors de l'attribution d'un numéro d'exemplaire : ${err}`);
+    //     });
+    // }
+    // updateBarcode();
 
-      io.on('append data', data => {
-        console.log(JSON.stringify(data, null, 2));
-        DBquery(io, 'INSERT INTO', data.table, {
-          text: `INSERT INTO ${data.table}(applicant_name, applicant_firstname, request_date, comment, status) VALUES($1, $2, $3, $4, $5)`,
-          values: data.values
-        });
+    io.on('append data', data => {
+      console.log(JSON.stringify(data, null, 2));
+      DBquery(io, 'INSERT INTO', data.table, {
+        text: `INSERT INTO ${data.table}(applicant_name, applicant_firstname, request_date, location_fk, comment, status) VALUES($1, $2, $3, $4, $5, $6)`,
+        values: data.values
       });
+    });
+  });
 
-      check4updates(io, tag);
+  check4updates(io, tag);
 
-      restart(io);
+  restart(io);
 
-      shutdown(io);
+  shutdown(io);
 
-      io.on('delete data', data => {
-        if (data.table === 'in_requests') {
-          let barcode = '';
+  io.on('delete data', data => {
+    if (data.table === 'in_requests') {
+      let barcode = '';
 
-          DBquery(io, 'SELECT', data.table, {
-              text: `SELECT barcode from ${data.table} WHERE pib_number = ${data.data}`
+      DBquery(io, 'SELECT', data.table, {
+          text: `SELECT barcode from ${data.table} WHERE pib_number = ${data.data}`
+        })
+        .then(res => {
+          barcode = res.rows[0].barcode;
+
+          DBquery(io, 'UPDATE', 'barcodes', {
+              text: `UPDATE barcodes SET available = true WHERE barcode ILIKE '${barcode}'`
             })
-            .then(res => {
-              barcode = res.rows[0].barcode;
-
-              DBquery(io, 'UPDATE', 'barcodes', {
-                  text: `UPDATE barcodes SET available = true WHERE barcode ILIKE '${barcode}'`
-                })
-                .catch(err => console.error(`Erreur lors de la mise à jour de la table barcodes : ${err}`));
-            })
-            .then(() => {
-              DBquery(io, 'DELETE FROM', data.table, {
-                  text: `DELETE FROM ${data.table} WHERE pib_number = '${data.data}'`
-                })
-                .catch(err => console.error(`L'emprunt ${data.data} n'a pas pu être supprimé : ${err}`));
-            })
-            .then(() => {
-              updateBarcode();
-            })
-            .catch(err => console.error(`Une erreur est survenue lors du processus de suppression de la demande ${data.data} : ${err}`))
-        } else if (data.table === 'out_requests') {
+            .catch(err => console.error(`Erreur lors de la mise à jour de la table barcodes : ${err}`));
+        })
+        .then(() => {
           DBquery(io, 'DELETE FROM', data.table, {
               text: `DELETE FROM ${data.table} WHERE pib_number = '${data.data}'`
             })
-            .catch(err => console.error(`Une erreur est survenue lors de la suppression du prêt ${data.data} : ${err}`));
-        }
-      });
+            .catch(err => console.error(`L'emprunt ${data.data} n'a pas pu être supprimé : ${err}`));
+        })
+        .then(() => {
+          updateBarcode();
+        })
+        .catch(err => console.error(`Une erreur est survenue lors du processus de suppression de la demande ${data.data} : ${err}`))
+    } else if (data.table === 'out_requests') {
+      DBquery(io, 'DELETE FROM', data.table, {
+          text: `DELETE FROM ${data.table} WHERE pib_number = '${data.data}'`
+        })
+        .catch(err => console.error(`Une erreur est survenue lors de la suppression du prêt ${data.data} : ${err}`));
+    }
+  });
 
-      io.on('export db', format => {
-        emptyDir('exports');
+  io.on('export db', format => {
+    emptyDir('exports');
 
-        if (format === 'csv') {
-          DBquery(io, 'COPY', 'in_requests', {
-              text: `COPY in_requests TO '${path.join(__dirname + '/exports/loans.csv')}' DELIMITER ',' CSV HEADER`
+    if (format === 'csv') {
+      DBquery(io, 'COPY', 'in_requests', {
+          text: `COPY in_requests TO '${path.join(__dirname + '/exports/loans.csv')}' DELIMITER ',' CSV HEADER`
+        })
+        .then(() => {
+          DBquery(io, 'COPY', 'out_requests', {
+              text: `COPY out_requests TO '${path.join(__dirname + '/exports/borrowings.csv')}' DELIMITER ',' CSV HEADER`
             })
             .then(() => {
-              DBquery(io, 'COPY', 'out_requests', {
-                  text: `COPY out_requests TO '${path.join(__dirname + '/exports/borrowings.csv')}' DELIMITER ',' CSV HEADER`
+              DBquery(io, 'COPY', 'drafts', {
+                  text: `COPY drafts TO '${path.join(__dirname + '/exports/drafts.csv')}' DELIMITER ',' CSV HEADER`
                 })
                 .then(() => {
-                  DBquery(io, 'COPY', 'drafts', {
-                      text: `COPY drafts TO '${path.join(__dirname + '/exports/drafts.csv')}' DELIMITER ',' CSV HEADER`
-                    })
-                    .then(() => {
-                      // Zip the received files before sending them to the client
-                      let zip = new admZip();
+                  // Zip the received files before sending them to the client
+                  let zip = new admZip();
 
-                      zip.addLocalFolder('exports', 'pib-manager-export.zip');
+                  zip.addLocalFolder('exports', 'pib-manager-export.zip');
 
-                      file2download.path = 'exports/pib-manager-export.zip';
-                      file2download.name = 'pib-manager-export.zip';
+                  file2download.path = 'exports/pib-manager-export.zip';
+                  file2download.name = 'pib-manager-export.zip';
 
-                      zip.writeZip(file2download.path);
-                      io.emit('export successfull');
-                    })
-                    .catch(err => {
-                      console.error(`Une erreur est survenue lors de l'export de la table des requêtes express : ${err}`);
-                    });
+                  zip.writeZip(file2download.path);
+                  io.emit('export successfull');
                 })
                 .catch(err => {
-                  console.error(`Une erreur est survenue lors de l'export de la table des prêts : ${err}`);
+                  console.error(`Une erreur est survenue lors de l'export de la table des requêtes express : ${err}`);
                 });
             })
             .catch(err => {
-              console.error(`Une erreur est survenue lors de l'export de la table des emprunts : ${err}`);
+              console.error(`Une erreur est survenue lors de l'export de la table des prêts : ${err}`);
             });
-        } else if (format === 'pgsql') {
-          file2download.path = 'exports/pib.pgsql';
-          file2download.name = 'pib.pgsql';
-          exportDB(file2download.path);
-
-          io.emit('export successfull');
-        }
-      });
-
-      io.on('mail request', reader => {
-        DBquery(io, 'SELECT', 'readers', {
-            text: `SELECT email, gender FROM readers WHERE name ILIKE '${reader}'`
-          })
-          .then(res => {
-            io.emit('mail retrieved', {
-              mail: res.rows[0].email.substring(7, 100),
-              gender: res.rows[0].gender
-            });
-          })
-          .catch(err => {
-            console.log(JSON.stringify(err, null, 2));
-          });
-      });
-
-      io.on('send mail', receiver => {
-        mail(receiver, client);
-        notify(io, 'mail');
-      });
-
-      io.on('retrieve readers', name => {
-        if (name.length >= 3) {
-          client.query(`SELECT name FROM readers WHERE name ILIKE '${name}%' LIMIT 5`)
-            .then(res => {
-              io.emit('readers retrieved', res.rows);
-            })
-            .catch(err => {
-              console.error(err);
-            });
-        }
-      });
-
-      io.on('settings', settings => {
-        if (settings.mail_content !== undefined) {
-          query = `UPDATE settings SET mail_content = '${settings.mail_content}'`;
-          console.log(query);
-        }
-
-        DBquery(io, 'UPDATE', 'settings', {
-          text: query
+        })
+        .catch(err => {
+          console.error(`Une erreur est survenue lors de l'export de la table des emprunts : ${err}`);
         });
-      });
-    });
-  })
+    } else if (format === 'pgsql') {
+      file2download.path = 'exports/pib.pgsql';
+      file2download.name = 'pib.pgsql';
+      exportDB(file2download.path);
 
-  .get('/search', (req, res) => {
+      io.emit('export successfull');
+    }
+  });
+
+  io.on('mail request', reader => {
+    DBquery(io, 'SELECT', 'readers', {
+        text: `SELECT email, gender FROM readers WHERE name ILIKE '${reader}'`
+      })
+      .then(res => {
+        io.emit('mail retrieved', {
+          mail: res.rows[0].email.substring(7, 100),
+          gender: res.rows[0].gender
+        });
+      })
+      .catch(err => {
+        console.log(JSON.stringify(err, null, 2));
+      });
+  });
+
+  io.on('send mail', receiver => {
+    mail(receiver, client);
+    notify(io, 'mail');
+  });
+
+  io.on('retrieve readers', name => {
+    if (name.length >= 3) {
+      client.query(`SELECT name FROM readers WHERE name ILIKE '${name}%' LIMIT 5`)
+        .then(res => {
+          io.emit('readers retrieved', res.rows);
+        })
+        .catch(err => {
+          console.error(err);
+        });
+    }
+  });
+
+  io.on('settings', settings => {
+    if (settings.mail_content !== undefined) {
+      query = `UPDATE settings SET mail_content = '${settings.mail_content}'`;
+      console.log(query);
+    }
+
+    DBquery(io, 'UPDATE', 'settings', {
+      text: query
+    });
+  });
+})
+
+.get('/search', (req, res) => {
     res.render('search.ejs', {
       currentVersion: tag,
       locations: process.env.LOCATIONS.split(','),
@@ -370,14 +368,15 @@ app.get('/', (req, res) => {
 
     io.once('connection', io => {
       io.on('search', data => {
-        if (!data.getApplicant) query = `SELECT * FROM ${data.table}`;
-        else query = `SELECT * FROM ${data.table} WHERE applicant_name ILIKE '%${data.applicant_name}%'`;
+        if (!data.getApplicant) query = `SELECT * FROM tasks WHERE location_fk = ${data.location}`;
+        else query = `SELECT * FROM tasks WHERE applicant_name ILIKE '%${data.applicant_name}%'`;
 
-        DBquery(io, 'SELECT', data.table, {
+        DBquery(io, 'SELECT', 'tasks', {
             text: query
           })
           .then(res => {
             if (res.rowCount !== 0 || res.rowCount !== null) {
+              console.log(JSON.stringify(res.rows, null, 2));
               io.emit('search results', res.rows);
             }
           });
@@ -399,11 +398,13 @@ app.get('/', (req, res) => {
       });
 
       io.on('delete data', data => {
-        query = `DELETE FROM tasks WHERE id = '${data.key}'`;
+        if (data.key !== undefined) {
+          query = `DELETE FROM tasks WHERE id = '${data.key}'`;
 
-        DBquery(io, 'DELETE FROM', data.table, {
-          text: query
-        });
+          DBquery(io, 'DELETE FROM', data.table, {
+            text: query
+          });
+        }
       });
     });
   })
